backend/pyproject.toml
[tool.poetry]
name = "construction-agent-poc"
version = "0.1.0"
description = "Simple conversational agent for Aurora construction documents"
authors = ["Construction Agent Team"]
readme = "../README.md"
packages = [{include = "*.py"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-dotenv = "^1.0.0"
openai = "^1.12.0"
websockets = "^12.0"
python-multipart = "^0.0.6"
pydantic = "^2.5.0"
httpx = "^0.25.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.12.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn app:app --reload --host 0.0.0.0 --port 8000"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ['py311']